+++ asynctor/compat.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+
+import os
+import sys
+
+if sys.version_info >= (3, 11):
+    from contextlib import chdir
+else:
+    import contextlib
+
+    class chdir(contextlib.AbstractContextManager):  # Copied from source code of Python3.13
+        """Non thread-safe context manager to change the current working directory."""
+
+        def __init__(self, path):
+            self.path = path
+            self._old_cwd = []
+
+        def __enter__(self):
+            self._old_cwd.append(os.getcwd())
+            os.chdir(self.path)
+
+        def __exit__(self, *excinfo):
+            os.chdir(self._old_cwd.pop())