+++ tests/test_compat.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+
+def test_chdir(tmp_work_dir):
+    assert tmp_work_dir == Path.cwd()